Extract parse_time to a separate function
Bug: b/542342091
