+++ base/cvd/cuttlefish/host/libs/web/BUILD.bazel
@@ -33,6 +33,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/web:android_build_url",
         "//cuttlefish/host/libs/web:build_api",
         "//cuttlefish/host/libs/web:credential_source",
+        "//cuttlefish/host/libs/web:parse_time",
         "//cuttlefish/host/libs/web/cas:cas_downloader",
         "//cuttlefish/host/libs/web/http_client",
         "//cuttlefish/host/libs/web/http_client:http_file",
@@ -235,5 +236,15 @@ cf_cc_library(
         "@fmt",
         "@jsoncpp",
         "@zlib",
+    ],
+)
+
+cf_cc_library(
+    name = "parse_time",
+    srcs = ["parse_time.cc"],
+    hdrs = ["parse_time.h"],
+    deps = [
+        "//cuttlefish/result:expect",
+        "//cuttlefish/result:result_type",
     ],
 )
